net.Dial() and return any error in sendZookeeperCmd
With zookeeper, we can't run multiple commands without a new
connection. With this commit, connection logic abstracted behind
sendZookeeperCmd and from now on it is responsible for creating
and closing connections. Also, handling command errors and parsing
the response logics are also abstracted behind sendZookeeperCmd.
So, one can use sendZookeeperCmd and they can know if error is
nil, first return value is a string list of response lines.